Lezen van json is volledig

+++ Main/src/Job.java
@@ -1,12 +1,12 @@
 public class Job {
-    private final long id;
-    private final long duration;
-    private final long releaseDate;
-    private final long dueDate;
-    private final double earlinessPenalty;
-    private final double rejectionPenalty;
+    private final int id;
+    private final int duration;
+    private final int releaseDate;
+    private final int dueDate;
+    private final float earlinessPenalty;
+    private final float rejectionPenalty;
     private long start=0, stop=0;
-    public Job(long id, long duration, long releaseDate, long dueDate, double earlinessPenalty, double rejectionPenalty){
+    public Job(int id, int duration, int releaseDate, int dueDate, float earlinessPenalty, float rejectionPenalty){
         this.id = id;
         this.duration = duration;
         this.releaseDate = releaseDate;
@@ -15,15 +15,15 @@ public class Job {
         this.rejectionPenalty = rejectionPenalty;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setStart(long start) {
+    public void setStart(int start) {
         this.start = start;
     }
 
-    public void setStop(long stop) {
+    public void setStop(int stop) {
         this.stop = stop;
     }
